#2634 Separate the code ina helper
#2634 Make the CalloutBankStatement webui compliant
